Move more Kotlin render and thread tests into commonTest
Reuse the owned-texture fixture for session lifecycle, readback, cluster
queries, and dummy set-target checks. Share a background-thread helper so
owner-thread and projection tests run on every target.
