+++ bindings/kotlin/src/androidDeviceTest/kotlin/org/maplibre/nativeffi/BackgroundThread.android.kt
@@ -1,0 +1,7 @@
+package org.maplibre.nativeffi
+
+internal actual fun runOnBackgroundThread(block: () -> Unit) {
+  val thread = Thread(block, "maplibre-test-background")
+  thread.start()
+  thread.join()
+}
