+++ bindings/kotlin/src/androidNativeTest/kotlin/org/maplibre/nativeffi/render/OwnedTextureTestSupport.androidNative.kt
@@ -1,0 +1,11 @@
+package org.maplibre.nativeffi.render
+
+import org.maplibre.nativeffi.map.MapHandle
+
+/**
+ * Kotlin/Native Android targets only compile. The Android device suite attaches an EGL owned
+ * texture for the common render tests.
+ */
+internal actual object OwnedTextureTestSupport {
+  actual fun attach(map: MapHandle, width: Int, height: Int): OwnedTextureTestSession? = null
+}
