+++ bindings/kotlin/src/commonTest/kotlin/org/maplibre/nativeffi/map/MapProjectionHandleTest.kt
@@ -1,17 +1,22 @@
 package org.maplibre.nativeffi.map
 
+import kotlin.concurrent.atomics.AtomicReference
+import kotlin.concurrent.atomics.ExperimentalAtomicApi
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertFalse
+import kotlin.test.assertNull
 import kotlin.test.assertTrue
 import org.maplibre.nativeffi.Maplibre
 import org.maplibre.nativeffi.camera.CameraOptions
 import org.maplibre.nativeffi.camera.EdgeInsets
 import org.maplibre.nativeffi.error.InvalidStateException
 import org.maplibre.nativeffi.geo.LatLng
+import org.maplibre.nativeffi.runOnBackgroundThread
 import org.maplibre.nativeffi.runtime.RuntimeHandle
 
+@OptIn(ExperimentalAtomicApi::class)
 class MapProjectionHandleTest {
   // BND-043, BND-103.
 
@@ -68,5 +73,26 @@ class MapProjectionHandleTest {
     } finally {
       runtime.close()
     }
+  }
+
+  @Test
+  fun projectionRemainsUsableOnAnotherThreadAfterMapClose() {
+    val runtime = RuntimeHandle.create(org.maplibre.nativeffi.runtime.RuntimeOptions())
+    val map = MapHandle.create(runtime, MapOptions())
+    val projection = map.createProjection()
+    map.close()
+    runtime.close()
+    val failure = AtomicReference<Throwable?>(null)
+
+    runOnBackgroundThread {
+      try {
+        projection.camera
+        projection.close()
+      } catch (error: Throwable) {
+        failure.store(error)
+      }
+    }
+
+    assertNull(failure.load())
   }
 }
