+++ bindings/kotlin/src/commonTest/kotlin/org/maplibre/nativeffi/runtime/OwnerThreadTest.kt
@@ -1,16 +1,15 @@
 package org.maplibre.nativeffi.runtime
 
-import java.util.concurrent.atomic.AtomicReference
-import kotlin.concurrent.thread
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 import org.maplibre.nativeffi.error.MaplibreStatus
 import org.maplibre.nativeffi.error.WrongThreadException
+import org.maplibre.nativeffi.failureFromBackgroundThread
 import org.maplibre.nativeffi.map.MapHandle
 import org.maplibre.nativeffi.map.MapOptions
 
-class OwnerThreadAndroidTest {
+class OwnerThreadTest {
   @Test
   fun runtimeAndMapCallsFromAnotherHostThreadCopyTheirDiagnostics() {
     val runtime = RuntimeHandle.create(RuntimeOptions())
@@ -23,13 +22,13 @@ class OwnerThreadAndroidTest {
         },
       )
     try {
-      val runtimeFailure = failureFromThread { runtime.pump(0) }
+      val runtimeFailure = failureFromBackgroundThread { runtime.pump(0) }
       val runtimeWrongThread = runtimeFailure as? WrongThreadException ?: throw runtimeFailure
       assertEquals(MaplibreStatus.WRONG_THREAD, runtimeWrongThread.status)
       val runtimeDiagnostic = runtimeWrongThread.diagnostic
       assertTrue(runtimeDiagnostic.isNotBlank())
 
-      val mapFailure = failureFromThread {
+      val mapFailure = failureFromBackgroundThread {
         map.setStyleJson("{\"version\":8,\"sources\":{},\"layers\":[]}".encodeToByteArray())
       }
       val mapWrongThread = mapFailure as? WrongThreadException ?: throw mapFailure
@@ -44,11 +43,5 @@ class OwnerThreadAndroidTest {
       map.close()
       runtime.close()
     }
-  }
-
-  private fun failureFromThread(block: () -> Unit): Throwable {
-    val result = AtomicReference<Throwable?>()
-    thread { result.set(runCatching(block).exceptionOrNull()) }.join()
-    return result.get() ?: error("wrong-thread operation succeeded")
   }
 }
